Late Latching Stage 3 Changes
Full Changes for Late Latching Stage 3, mostly consist of the mark/unmark calls, as well as the generic BeginRecording/EndRecording. Mark and unmark is carefully done for main rendering of the forward passes(not the prepass or post-passes). This way, we can be sure this works with any type of developer use case, with heavy post-processing. This was tested thoroughly with the URP template scene.